fix(extractor): Use output_dim parameter instead of hard-coded values
Previously, a hard-coded value of 256 was used, which relied on output_dim
being passed correctly as 256 by calling code.
The default value in the constructors even suggested that a value of 128 was used
which was misleading because you couldn't see the connection between output_dim
and the hard-coded values of 256.
So remove the default output_dim value and use output_dim in the function bodies.